+++ changelog.md
@@ -1,6 +1,6 @@
 ### 0.4.3
 #### Documentation
-- simplify and fix layout of ``count`` examples
+- simplify and fix layout of ``count``, ``cumulative`` and ``max`` examples
  
 ### 0.4.2
 #### CI changes