+++ changelog.md
@@ -1,11 +1,20 @@
-## 0.4.2
+### 0.4.3
+#### Added
+- `disjunctive` constraint
+#### Fixed
+- array slices with upper slice as operations should compile correctly now
+#### Documentation
+- simplify and fix layout of ``count``, ``cumulative`` 
+  ``table`` and ``max`` examples
+ 
+### 0.4.2
 #### CI changes
 - use minizinc 2.7.6 as maximum version in CI (as in minizinc-python)
 - use minizinc 2.6.0 as minimum version in CI (as in minizinc-python)
 #### Python interpreters support
 - add 3.12 CPython
 
-## 0.4.1
+### 0.4.1
 #### Added
 - ``table`` constraint
 - ``contains`` method for arrays and sets, to check if elem presented